app/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
sentence-transformers
openai
python-dotenv
